requirements.txt
boto3==1.9.77
botocore==1.12.77
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-dynamic-fixture==2.0.0
django-extensions==2.1.4
django-filter==2.0.0
django-filters==0.2.1
django-storages==1.7.1
djangorestframework==3.9.0
docutils==0.14
idna==2.8
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
Markdown==3.0.1
MarkupSafe==1.1.0
Pillow==5.4.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
Pygments==2.3.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
s3transfer==0.1.13
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
